Added Project and ProjectManager into apps.proyecto.models w/ UnitTests. Added UserStory and Notes into apps.administracion.models w/ UnitTests. Fixed apps.administracion.views bugs about model migrations
